requirements.txt
beautifulsoup4==4.12.3
langchain==0.1.16
python-dotenv==1.0.1
webdriver_manager==4.0.1
